// src/app/models/QuestionType.ts
/** Valeurs alignées sur l’enum côté Spring / base. */
export enum QuestionType {
  YES_OR_NO = 'YES_OR_NO',
  NUMBER = 'NUMBER',
  SHORT_ANSWER = 'SHORT_ANSWER',
  PARAGRAPH = 'PARAGRAPH',
  MULTIPLE_CHOICE = 'MULTIPLE_CHOICE',
  CHECKBOXES = 'CHECKBOXES',
  DROPDOWN = 'DROPDOWN',
  DATE = 'DATE',
  TIME = 'TIME',
}